feat: SAA 적용 및 Navigation 구조 적용
