+++ app/src/main/AndroidManifest.xml
@@ -17,25 +17,14 @@
             android:label="@string/title_activity_folder"
             android:theme="@style/Theme.LETSSOPT" />
         <activity
-            android:name=".presentation.SignUpActivity"
-            android:exported="false"
-            android:label="SignUpActivity"
-            android:theme="@style/Theme.LETSSOPT" />
-        <activity
-            android:name=".presentation.LoginActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true"
-            android:label="LoginActivity"
+            android:label="MainActivity"
             android:theme="@style/Theme.LETSSOPT">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
             <category android:name="android.intent.category.LAUNCHER" />
         </intent-filter>
-        </activity>
-        <activity
-            android:name=".presentation.MainActivity"
-            android:exported="false"
-            android:label="MainActivity"
-            android:theme="@style/Theme.LETSSOPT">
         </activity>
     </application>
