+++ app/src/main/java/com/example/letssopt/presentation/login/LoginScreen.kt
@@ -1,12 +1,6 @@
-package com.example.letssopt.presentation
+package com.example.letssopt.presentation.login
 
-import android.content.Intent
-import android.os.Bundle
 import android.widget.Toast
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -14,7 +8,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -23,9 +16,9 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.letssopt.R
 import com.example.letssopt.common.util.isLoginValid
 import com.example.letssopt.designsystem.component.Button.WatchaButton
@@ -33,67 +26,18 @@ import com.example.letssopt.designsystem.component.Text.WatchaFormField
 import com.example.letssopt.designsystem.component.Text.WatchaSemiTitle
 import com.example.letssopt.designsystem.component.Text.WatchaTextField
 import com.example.letssopt.designsystem.theme.Background
-import com.example.letssopt.designsystem.theme.LETSSOPTTheme
 import com.example.letssopt.designsystem.theme.PrimaryRed
 import com.example.letssopt.designsystem.theme.TextSecondary
-import com.example.letssopt.presentation.viewmodel.LoginViewModel
 
-class LoginActivity : ComponentActivity() {
-    private val viewModel: LoginViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        val sharedPreferences = getSharedPreferences("login_preferences", MODE_PRIVATE)
-
-        if (sharedPreferences.getBoolean("auto_login", false)) {
-            startActivity(Intent(this, MainActivity::class.java))
-            finish()
-            return
-        }
-
-        setContent {
-            LETSSOPTTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    LoginScreen(
-                        modifier = Modifier.padding(innerPadding),
-                        email = viewModel.email.value,
-                        password = viewModel.password.value,
-                        onEmailChange = viewModel::updateEmail,
-                        onPasswordChange = viewModel::updatePassword,
-                        savedEmail = intent.getStringExtra("email"),
-                        savedPassword = intent.getStringExtra("password"),
-                        toSignUp = {
-                            val intent = Intent(this, SignUpActivity::class.java)
-                            startActivity(intent)
-                        },
-                        toMain = {
-                            sharedPreferences.edit()
-                                .putBoolean("auto_login", true)
-                                .apply()
-
-                            val intent = Intent(this, MainActivity::class.java)
-                            startActivity(intent)
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 fun LoginScreen(
     modifier: Modifier = Modifier,
-    email: String,
-    password: String,
-    onEmailChange: (String) -> Unit,
-    onPasswordChange: (String) -> Unit,
     savedEmail: String?,
     savedPassword: String?,
     toSignUp: () -> Unit,
-    toMain: () -> Unit
+    toMain: () -> Unit,
+    viewModel: LoginViewModel = viewModel()
 ) {
     val context = LocalContext.current
 
@@ -130,8 +74,8 @@ fun LoginScreen(
         )
 
         WatchaTextField(
-            value = email,
-            onValueChange = onEmailChange,
+            value = viewModel.email.value,
+            onValueChange = viewModel::updateEmail,
             placeholder = "이메일 주소를 입력하세요",
         )
 
@@ -143,8 +87,8 @@ fun LoginScreen(
         )
 
         WatchaTextField(
-            value = password,
-            onValueChange = onPasswordChange,
+            value = viewModel.password.value,
+            onValueChange = viewModel::updatePassword,
             placeholder = "비밀번호를 입력하세요",
             isPassword = true
         )
@@ -166,7 +110,12 @@ fun LoginScreen(
             text = "로그인",
             modifier = Modifier.padding(bottom = 50.dp),
             onClick = {
-                val result = isLoginValid(email, password, savedEmail, savedPassword)
+                val result = isLoginValid(
+                    viewModel.email.value,
+                    viewModel.password.value,
+                    savedEmail,
+                    savedPassword
+                )
 
                 Toast.makeText(context, result, Toast.LENGTH_SHORT).show()
 
@@ -178,19 +127,19 @@ fun LoginScreen(
     }
 }
 
-@Preview(showBackground = true)
-@Composable
-private fun LoginScreenPreview() {
-    LETSSOPTTheme {
-        LoginScreen(
-            email = "test@email.com",
-            password = "12345678",
-            onEmailChange = {},
-            onPasswordChange = {},
-            savedEmail = "email",
-            savedPassword = "password",
-            toSignUp = {},
-            toMain = {}
-        )
-    }
-}
+//@Preview(showBackground = true)
+//@Composable
+//private fun LoginScreenPreview() {
+//    LETSSOPTTheme {
+//        LoginScreen(
+//            email = "test@email.com",
+//            password = "12345678",
+//            onEmailChange = {},
+//            onPasswordChange = {},
+//            savedEmail = "email",
+//            savedPassword = "password",
+//            toSignUp = {},
+//            toMain = {}
+//        )
+//    }
+//}
