+++ app/src/main/java/com/example/letssopt/presentation/login/LoginViewModel.kt
@@ -1,16 +1,13 @@
-package com.example.letssopt.presentation.viewmodel
+package com.example.letssopt.presentation.login
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 
-class SignUpViewModel : ViewModel() {
+class LoginViewModel : ViewModel() {
     var email = mutableStateOf("")
         private set
 
     var password = mutableStateOf("")
-        private set
-
-    var passwordCheck = mutableStateOf("")
         private set
 
     fun updateEmail(input: String) {
@@ -19,9 +16,5 @@ class SignUpViewModel : ViewModel() {
 
     fun updatePassword(input: String) {
         password.value = input
-    }
-
-    fun updatePasswordCheck(input: String) {
-        passwordCheck.value = input
     }
 }
