+++ app/src/main/java/com/example/letssopt/presentation/navigation/Route.kt
@@ -1,0 +1,12 @@
+package com.example.letssopt.presentation.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object Login
+
+@Serializable
+data object SignUp
+
+@Serializable
+data object Main
