+++ app/src/main/java/com/example/letssopt/presentation/signup/SignUpViewModel.kt
@@ -1,13 +1,16 @@
-package com.example.letssopt.presentation.viewmodel
+package com.example.letssopt.presentation.signup
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 
-class LoginViewModel : ViewModel() {
+class SignUpViewModel : ViewModel() {
     var email = mutableStateOf("")
         private set
 
     var password = mutableStateOf("")
+        private set
+
+    var passwordCheck = mutableStateOf("")
         private set
 
     fun updateEmail(input: String) {
@@ -16,5 +19,9 @@ class LoginViewModel : ViewModel() {
 
     fun updatePassword(input: String) {
         password.value = input
+    }
+
+    fun updatePasswordCheck(input: String) {
+        passwordCheck.value = input
     }
 }
